API

API

DICE.v2013R

DICE.v2016R

DICE.v2016R2

DICE.vCJL

DICE.v2013R — Method.

v2013R([flavour])

Identifier for the 2013R version of the model. Defaults to the Vanilla flavour, but also accepts RockyRoad.

Examples

julia> v2013R()
v2013R (Vanilla flavour)

julia> v2013R(Vanilla)
v2013R (Vanilla flavour)

julia> v2013R(RockyRoad)
v2013R (Rocky Road flavour)

source

DICE.v2016R — Method.

v2016R()

Identifier for the 2016R beta version of the model. Note that this is a beta version following DICE2016R-091916ap.gms.

Examples

julia> v2016R()
v2016R beta

source

DICE.v2016R2 — Method.

v2016R2()

Identifier for the 2016R2 version of the model, following the DICE2016R2-Unc-{base|opt}-v5-55555-083017B.gms file from November 2017, defaulting to the 'Best Guess' parameters.

Examples

julia> v2016R2()
v2016R2

source

DICE.vCJL — Method.

vCJL()

Identifier for the DICE-CJL version of the model.

Examples

julia> vCJL()
CJL

source